+++ src/lib/components/PomodoroEnd.svelte
@@ -4,8 +4,10 @@
 
     let {
         pomodoroEndScreen = $bindable(),
+        pomodoroDataSent = $bindable(),
+        completedTasks,
+        totalTasks,
         time,
-        xp = 0,
         blockType = "Grass"
     } = $props()
     
@@ -13,15 +15,18 @@
     
     function handleKeyPress(event: KeyboardEvent) {
         event.preventDefault()
-        if (event.key == "Escape") pomodoroEndScreen = false
+        if (event.key == "Escape") {
+            pomodoroDataSent = false
+            pomodoroEndScreen = false
+        }
     }
     
-    onMount(() => {
+    onMount(() => {       
         if (time <= 600) {
             message = pomEndMessages.bad
-        } else if (xp > 600 && xp <= 1200) {
+        } else if (time > 600 && time <= 1200) {
             message = pomEndMessages.ok
-        } else if (xp > 1200) {
+        } else if (time > 1200) {
             message = pomEndMessages.good
         }
     })
@@ -31,16 +36,19 @@
 
 <div
     class="bg-transparent cursor-pointer w-full h-full flex justify-center items-center"
-    onclick={() => pomodoroEndScreen = false}
+    onclick={() => {pomodoroDataSent = false; pomodoroEndScreen = false}}
 >
-    <div class="p-5 px-10 bg-white cursor-default w-fit max-w-150 h-fit rounded-xl flex flex-col items-center">
+    <div class="p-5 px-10 shadow-2xl backdrop-blur-xs cursor-default w-fit max-w-150 h-fit rounded-xl flex flex-col items-center">
         <div class="flex flex-row gap-4">
             <div class="flex-2">
                 <h2 class={time<600 ? 'text-red-500': ''}>{formatTime(time)}</h2>
                 <p>{message}</p>
+                {#if totalTasks > 0}
+                    <p>You Completed {completedTasks} out of {totalTasks} Tasks From Your List</p>
+                {/if}
             </div>
             <div class="flex-1 text-right">
-                <h3>You Gained a {blockType} Block</h3>
+                <h3 class="text-white text-shadow-lg font-bold ">You Gained a {blockType} Block</h3>
             </div>
         </div>
         <p class="m-5 cursor-pointer">Click anywhere to close or press esc</p>