+++ src/routes/+layout.svelte
@@ -1,6 +1,7 @@
-<script>
-  let { children } = $props();
-  import "../app.css";
+<script lang="ts">
+    let { children } = $props();
+    import "../app.css";
 </script>
 
+<svelte:window on:contextmenu={(e) => e.preventDefault()} />
 {@render children()}