+++ Changelog.md
@@ -2,13 +2,48 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.0.0] - 2025-11-16
+## [alpha-beta-1.1.0] - 2025-11-20
+
+### Added
+
+- Added icon to audio button that toggles
+- Version number to homescreen
+- Animation to block tooltip
+- To do list function for each pomodoro session
+- Tooltip now indicates to the user which block is their first block
+- Spotlight that syncs to time of day
+- Random rotation to trees, bushes & flowers
+- Loading screen on app launch
+- Rotation island view with horizontal trackpad movement
+- Breaks between pomodoro sessions
+
+## Changes
+
+- Reduced the chances of grass blocks appearing
+- Multiple blocks are now added depending on how long the user studied
+    - under 5 mins - grass/water block
+    - under 10 mins - grass/water block + flower block
+    - under 15 mins - 2x grass/water block + bush/flower block
+    - under 20 mins - 2x grass/water block + bush block
+    - under 25 mins - 3x grass/water block + tree block
+    - over 25 mins - 3x grass/water block + 2x bush/flower block + tree
+- Minor UI changes
+
+### Fixes
+
+- User can now pan the camera while in a pomodoro session
+- Pomodoro end screen now displays the 'ok' and 'good' messages
+- Tooltips now don't display after hovering on another block
+- XP resets when the user resets their data
+
+---
+
+## [alpha-1.0.0] - 2025-11-16
 
 ### Added
 
 - Better lighting & shadows
 - Block appear animation
-- To do list appears when timer starts
 - Changed grass texture
 - Ability to pause the timer
 - Enter key press now triggers timer to start
@@ -32,7 +67,7 @@ All notable changes to this project will be documented in this file.
 
 -README
 
-## [0.6.0] - 2025-11-15
+## [alpha-0.6.0] - 2025-11-15
 
 ### Added
 
@@ -45,7 +80,7 @@ All notable changes to this project will be documented in this file.
 - Can now override numbers in timer input
 - Numbers are only allowed in timer input
 
-## [0.6.0] - 2025-11-15
+## [alpha-0.6.0] - 2025-11-15
 
 ### Added
 
@@ -55,7 +90,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.5.1] - 2025-11-15
+## [alpha-0.5.1] - 2025-11-15
 
 ### Fixes
 
@@ -63,7 +98,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.5.0] - 2025-11-15
+## [alpha-0.5.0] - 2025-11-15
 
 ### Added
 
@@ -76,10 +111,11 @@ All notable changes to this project will be documented in this file.
 ### Changes
 
 - Lowered hances of a grass block appearing if the user only does a short pomodoro session
+- UI text colours
 
 ---
 
-## [0.4.0] - 2025-11-15
+## [alpha-0.4.0] - 2025-11-15
 
 ### Added
 
@@ -97,18 +133,21 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.3.0] - 2025-11-15
+## [alpha-0.3.0] - 2025-11-15
 
 ### Added
 
 - Xp now accumulates with each pomodoro session
 - New window when a pomodoro session ends
 - Made scene camera more sensitive to cursor
+- Zoom mechanic by scrolling
+- Random rotation to blocks
 
 ### Changes
 
 - Camera position
 - UI and scene colours
+- Other block tooltips dont appear when the user hovers over other blocks 
 
 ### Fixed
 
@@ -117,7 +156,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.2.0] - 2025-11-15
+## [alpha-0.2.0] - 2025-11-15
 
 ### Added
 
@@ -127,7 +166,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.1.0] - 2025-11-14
+## [alpha-0.1.0] - 2025-11-14
 
 ### Added
 
@@ -139,7 +178,7 @@ All notable changes to this project will be documented in this file.
 
 ---
 
-## [0.0.1] - 2025-11-14
+## [alpha-0.0.1] - 2025-11-14
 
 ### Added